+++ libc/src/lcmath.c
@@ -1,9 +1,9 @@
 #include "lcmath.h"
 
-int ceil(float n)
+int64_t ceil(double n)
 {
         if (n > 0)
-                return n - (int)n > 0 ? (int)(n + 1) : (int)n;
+                return n - (int64_t)n > 0 ? (int64_t)(n + 1) : (int64_t)n;
         else
-                return n - (int)n < 0 ? (int)(n - 1) : (int)n;
+                return n - (int64_t)n < 0 ? (int64_t)(n - 1) : (int64_t)n;
 }
